# asses/target/classes/api-docs.yaml
openapi: 3.0.1
servers:
- url: http://localhost:9090
  description: Generated server url
paths:
  /employee/{empid}:
    put:
      tags:
      - employee-controller
      operationId: updateEmployee
      parameters:
      - name: empid
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    delete:
      tags:
      - employee-controller
      operationId: deleteEmployeebyId
      parameters:
      - name: empid
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /employee:
    get:
      tags:
      - employee-controller
      operationId: getAllEmployee
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
    post:
      tags:
      - employee-controller
      operationId: createEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /search:
    get:
      tags:
      - employee-controller
      operationId: searchEmployee
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
components:
  schemas:
    Employee:
      type: object
      properties:
        empid:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string